third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-get-errors.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/serviceworker/resources/interfaces.js"></script>
<script>
function stubResolverUndefinedChecker(c) {
    assert_equals(c, undefined);
    this.done();
}

function stubRejectionChecker(reason) {
    assert_unreached("get() should not reject, but did: " + reason.name);
}

promise_test(function (t) {
    if (window.testRunner)
        testRunner.addMockCredentialManagerError("pending");
    return promise_rejects(t, "InvalidStateError", navigator.credentials.get({ password: true }));
});
promise_test(function (t) {
    if (window.testRunner)
        testRunner.addMockCredentialManagerError("disabled");
    return promise_rejects(t, "InvalidStateError", navigator.credentials.get({ password: true }));
});
promise_test(function (t) {
    if (window.testRunner)
        testRunner.addMockCredentialManagerError("unknown");
    return promise_rejects(t, "NotReadableError", navigator.credentials.get({ password: true }));
});
</script>
